feat : UserItem(소유아이템) 관련 API 기능 추가
- Game 관련 커스텀 예외 추가
- PlacedRoomItem 자료형 수정

+++ src/main/java/com/jhlab/ninety/domain/game/room/entity/PlacedRoomItem.java
@@ -26,7 +26,7 @@ public class PlacedRoomItem extends BaseEntity {
     @JoinColumn(name = "item_id")
     private RoomItem item;
 
-    private Integer posX;
-    private Integer posY;
-    private Integer rotation;
+    private double posX;
+    private double posY;
+    private double rotation;
 }
